backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
pydantic[email]>=2.6.0
pydantic-settings>=2.2.0
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
redis>=5.0.0
minio>=7.2.0
celery>=5.3.6
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.1
python-magic>=0.4.27
